// src/app/interfaces/sysrecord.ts
export interface Hospital {
    hospitalCode: string;
    hospitalName: string;
    hospitalShortName: string;
}
// ...
export interface Register {
    userName: string;
    email: string;
    password: string;
    employeeCode: string;
    hospitalCode: string;
    prefixName: string;
    firstName: string;
    lastName: string;
    isAdmin: boolean;
    createBy: string;
    langCode: string;
}

export interface GetHospitalListByEmp {
    hospitalCode: string;
    hospitalName: string;
    hospitalShortName: string;
}

export interface GetEmployeeProfile {
    hospitalCode: string;
    employeeCode: string;
    prefixName: string;
    firstName: string;
    lastName: string;
    langCode: string;
}